[DAPHNE-#78] Extended matrix multiplication kernel (dgemm/dgemv/ddot)
So far the matrix multiplication kernel always called sgemm/dgemm, even
for matrix-vector and vector-vector operations which was unnecessary
inefficient. Normally, these operations are memory-bandwidth-bound,
but for small intermediates that fit in caches an internal
dispatching improves performance. Additionally this patch also
includes new matrix-vector tests.

As a next step we should also extend the APIs for multi-threaded
operations.

Closes #78.