[AMLS] Add distributed computation prototype
This adds basic distributed operations to the daphne dialect. Kernels are there, even though they are limited to `DenseMatrix<double>`. Since string representations are still not fully implemented I am always sending the same mlir code fragment to the workers, **THIS IS OF COURSE SEMENTICALLY INCORRECT**.

A distributed `Handle` holds row-column-index to distributed-data-reference mappings. During computation I plan (currently only data that is on the same worker is used) sending the same row-column-indexes of the inputs to a worker and operating on that. This does not work with matrix multiply, therefore a more general approach for the key of the mapping has to be used. How the keys are mapped is also up to discussion, I wrote two ideas as a todo in the TableGen of the Operations.

An idea for a more general approach of the key would be to use a variable length of indices, so for matrix multiply `A @ B` we would use `ROW_OF_A`x`COMMON_DIM`x`COLUMN_OF_B`, and just repeat the distributed-data-references for `A` and `B`. This should also map nicely to tensors, although with overhead that could be optimized away by an appropriate data structure for the `HandleMap`.

Also note the current `BLOCK_SIZE` is almost at the limit of the maximum protobuf/gRPC message length.