[DAPHNE-#106] Vectorized execution engine II (scheduler, worker, tasks)
This patch extends the vectorized execution engine by all necessary
means for multi-threaded operations with two inputs and a single output
that can be parallelized with pure row partitioning. In detail, this
includes a wrapper for managing threads, queues, creating tasks, and
waiting for completion, as well as basic CPU workers and task
extensions. Furthermore, the dense matrix has been extended by simple
slicing operations for zero-copy extraction of sub-matrices.