[DAPHNE-XX] Improved CUDA ops performance by keeping intermediates around
This commit introduces the method getValuesCUDA() to the DenseMatrix and changes the behavior of getValues() slightly. This is a preliminary version of "memory management" to keep track of whether there is a CUDA buffer and if the data is there or not (dirty flags). Furthermore, this commit contains a few cleanups and fixes (affine matmult, kernel instantiations).