[DAPHNE-#99] Script for distributed workers
A first bash script which connects to remote machines with ssh and can:

* build and deploy DistributedWorkers to remote machines
* start workers
* check status of running workers
* kill workers

Workers' own IPs and ports to listen to, can be specified inside the script, or with `--peers [IP[:PORT]],[IP[:PORT]],...`. Default port for all workers is 50000, but this can also be specified inside the script or with `-p,--port PORT`. If running on same machine (e.g. localhost), different ports must be specified.

With `--deploy` the script builds `DistributedWorker` executable (`./build.sh --target DistributedWorker`), compresses the `build` folder and uses `scp` and `ssh` to send and decompress at remote machines, inside the directory specified by `--pathToBuild` (default `~/DaphneDistributedWorker/`). If running workers on localhost, `PATH_TO_BUILD` can be set `/path/to/prototype` and provided `DistributedWorker` is built, `--deploy` is not nessecary.

Ssh username must be specified inside the script. For now the script assumes all remote machines can be accessed with the same `username`, `id_rsa` key and ssh port (default 22).

Note: given the complexity of this script, a Python script might be a better approach.

Note 2: a nice addition to this script might be to include: `export DISTRIBUTED_WORKERS=[peers]` after parsing peers, but for now the user must set the variable (before running `./build/bin/daphnec script.daphne`).

Usage example:

```
./deployDistributed.sh --deploy --pathToBuild /path/to/dir --peers localhost:5000,localhost:5001
./deployDistributed.sh -r # (Uses default peers and path/to/build/ to start workers)
```