Improved vectorized fusion.
This PR supports 2 return values and greatly improves fusion capabilities (important for the lm example in the CIDR paper, although this has to be inspected more, because the matrix multiply is currently not fused as matmul just slices the bigger matrix and keeps the other one whole as pipeline input).

Also lays the groundwork for \* return values of pipelines, even though we still have to inspect how we can handle varying return types.