[DAPHNE-183] Compiler pass to mark ops for CUDA execution
This patch evaluates implementation availability, minimum (matrix) input dimensions and max mem requirements (considering available device memory) to mark operations for CUDA execution.
The RewriteToCallKernelOpPass will add "CUDA_" to the name of the kernel to be called if the operation has its cuda_device attribute set to a valid ID (>=0).

Closes #183