This is a commit of branch `236-copying-deploy-files-to-merge` for Issue No. 236.
The branch `daphne-Vega` contains a set of **scripts** and **experimental data** files regarding:
- deployment of the `daphne` executable,
- packaging/virtualization of the installation, and
- collection and analysis of experimental data obtained through running this deployment.
It builds, packages, and deploys the code. It also includes a demo.

This file set from the branch `daphne-Vega` is now split to a smaller subset in order to be merged in the `prototype` GitLab repository's main branch through the branch `236-copying-deploy-files-to-merge`.
Therefore, a new branch was created for such first split.

The current files for movement within this branch (initially, Issue 236) are the files currently already existing at branch `daphne-Vega`:
1. **deployment**-related files:
- `Dockerfile`,
- `deployDistributed.sh`, and
2. **virtualization configuration** _scripts_ for building the **Singularity image** and running such an image:
- `create-d-sif.sh`,
- `e.daphne`, and
- `run.sh`.
These files resided in the directory `deploy/daphne-container/container/` of branch `daphne-Vega`.
These files are now copied and appropriately named within this branch, `236-copying-deploy-files-to-merge`:
- `create-d-sif.sh` -> `build-daphne-singularity-image.sh`,
- `e.daphne` -> `example-time.daphne`,
- `run.sh` -> `deploy-distributed-on-slurm.sh`, and
- `singularity`.

Due to the size of the Singularity image, a built image file itself is not included in the source tree within this commit.
Some more usage code for the deploy-distributed-on-slurm.sh script needs to be added.