[DAPHNE-#243] Add principal component analysis (PCA) algorithm script
This patch adds a basic version of PCA to the set of provided algorithm
scripts. At it's core this script performs an eigen factorization of
the covariance matrix. The integrated of Eigen into the parser as well
as missing operations (order on frames / or cast to frame) and table
require additional work though.