[DAPHNE-428] Distributed Runtime Bugfixes & Cleanups
This commit merges a longer development process to the main branch. The general topic is given in the first line of the commit message and the aggregated individual commit messages are listed below.

Closes #428

- Fixed various memory leaks.
- Fixed headers.
- Changed some pass-by-value to const-reference.
- Channel map changed to inline static.

[MINOR] Silenced some warnings

[Bugfix] Worker receiving and storing a value.

When the worker receives a value, we need to allocate memory, store the value and keep the address, for later use.
Since we did not allocate any memory this resulted in a bug where the address stored was not pointing to any value.

[Bugfix] Distribute/Broadcast kernels never actually checked if something was already placed at the workers.