[DAPHNE-#416] Initial Intel FPGA Support
This commit introduces support for the Intel PAC D5005 FPGA accelerator card. It adds the necessary bits in the compiler infrastructure, a runtime kernel for SGEMM and context handling. The precompiled bitstream will be provided in the separate repository github.com/daphne-eu/supplemental-binaries to avoid blowing up the size of the main repository.