[DAPHNE-451] FPGA GEMV Operation
This commit adds support for a single precision GEMV operation to the list of supported FPGA operations. The necessary bitstream can be downloaded from the supplemental-binaries git repository (https://github.com/daphne-eu/supplemental-binaries). This does not represent a separate DAPHNE kernel but branches out to call the appropriate function in the Matmult kernel.