[MINOR] Install MLIR in installPrefix
This change adds the invocation of the installation target while compiling LLVM/MLIR. With that, we can omit specifying LLVM_DIR and MLIR_DIR explicitly. Furthermore, this cuts the dependency of buildPrefix, which might not contain a meaningful value in a container with prebuild 3rd party dependencies. Finally, this enables us to switch the directory of installed dependencies with the installPrefix parameter of build.sh.