[BUGFIX] Use the static versions of Arrow and Parquet libs
With our current build system, if we use a shared thirdparty library,
the resulting binary (executable or shared library) depends on the
absolute path of the shared library. This means that it might not be
possible to run Daphne on a host other than the one we built it on.