[DAPHNE-#503] Sharing meta data among slices
When slicing a matrix, the MetaDataObject(MDO) is now shared with the shallow copy that is created in the slice operation.
This fixes a bug when, for example, printing a subset of a matrix after it was *initially* created on GPU (e.g., the result of a CUDA operation like matmult), where no memory was allocated on the host side (the values shared_ptr). In this partial fix, the memory needed to hold all data of the original matrix is allocated so that the sliced matrix can create a pointer+offset into it.
A full solution would put the values shared_ptr behind the MDO mechanism.

Closes #503