Added load partioning for distributed runtime.
Abstracted duplicate code for different distributed runtime communication frameworks.
Added LoadPartitioningDistributed.h class which is used to:

Split ranges based on some policy. For now only broadcast and simple distribution is supported.
Sets the metadata of the matrix being distributed/broadcasted.
Contains a static method which sets the metadata of the outputs during the computation step.
This class can be extended in order to support more complex distribution strategies.