[DAPHNE-#533] Bugfix: Wrong indexing in matrix views
In some cases when dealing with views on existing matrices, the data storage array would not be of the expected size (e.g., copying back from GPU in a view that is smaller than the original matrix would allocate the smaller amount of memory according to the view and copy back the required amount of data without the correct offset (which is actually a separate problem)). This commit should fix this by tracking if matrix dimensions and buffer sizes match up and taking appropriate measures for pos() calculation and offsets if needed. This concerns DenseMatrix only for now.

Closes #533