[BUGFIX, DAPHNE-#555] Catching SIG_ABRT due to unhandled exceptions
On ARM, unhandled exceptions from a library seem to cause an abort(), which is something we can handle by registering a handler for this abort posix signal (SIG_ABRT). By doing so we can at least pass the information that something happened without a major refactoring to introduce return code handling from the kernel libraries.

Closes #555