[BUGFIX, DAPHNE-#555] Catching more exceptions in compiler passes
This is a follow up fix for [1] to not only handle uncaught exceptions that come in via SIG_ABRT from the kernel libraries, but also from exceptions thrown in the DAPHNE compiler passes. Separate treatment is required here as a few script level test cases expect the DAPHNE executable to return PASS_ERROR if something fails in the compiler passes.

[1] commit 4a8f003cf5b499bec4ecbb138441af8dfa2413f0