[BUGFIX] Type cast while parsing transpose crashes on unknown input
The builder defined for the transpose operation in DaphneOps.td used a cast to MatrixType. This caused an assertion to fire while running a script (PCA) in DAPHNE debug that used a transpose operation with input from ctable().
The cast in the builder definition would generate code that is run while creating the transpose op in the parser stage of the DAPHNE compiler. At this stage the output of ctable() is still unknown because the type inference pass did not run yet.
Removing the cast is a viable solution as the subsequent type inference will set the types right when it is time to.