[DAPHNE-#575] Improved error logging with default logger; cleanups
This commit improves the behavior of the default logger to properly log all errors from the compiler passes and runtime kernels and test cases.
An initial effort has been made to have more uniform error messages but this remains work in progress.
A few occurrences of std::cout and std::cerr are left for the original authors of that code to decide how to log these outputs.

Closes #575