[DAPHNE-#582] Fix passing str variable to readFrame/readMatrix
This commit fixes a crash when passing "dynamic strings" that have not yet been materialized in the paring stage to the getMetaData() function. The current solution is to move this functionality out of the parsing stage and into the type inference stage of the DAPHNE compiler.

Incorporated review suggestions

- Implement DaphneInferFrameLabelsOpInterface for ReadOp
- Added comment lines to test scripts
- Fixed IsConstant Helper function for string types
- Added further example for dynamic path computation and fixed null string problem - still does not work since the string cannot be (yet) computed at compile time

Fixes daphne-eu/daphne#582
Closes daphne-eu/daphne#598