[DAPHNE-#560] Parsing unary minus operator
This change decouples the '-' from INT_LITERAL and FLOAT_LITERAL and handles a unary minus separately.

The test case covers a collection of cases where a unary minus is used and also documents where it still does not work.

This change also makes the command line argument an expression, so simple mathematical expressions like 2+2 will be evaluated.

Resolves #560