Fix MPI + removed unecessary comments.
- DistributedPipelines are executed by using Distributed kernels (i.e.
Distribute.h, Collect.h etc.). These kernels should be implemented as
stand-alone (so that the compiler can directly use them in the future if
needs to). MPI implementation was chaining execution from
DistributedCompute.h kernel to DistributedCollect.h, sending messages
from compute step and waiting for responses directly
after (at collect step). This approach works for pipelines but if
the compiler ever needs to single-use Distributedcollect.h,
the kernel should be callable as a
stand-alone kernel. This is now fixed.
- Removed unnecessary commented-out debug messages.