Distributed Worker: Merge arguments into IR.
- Distributed workers receive an IR code that needs to be computed. It
is received as a function that the worker currently pass arguments.
However this causes issues since some optimization passes require some
arguments to be constant (which they are, just not declared in the IR).
With this commit, workers, modify the IR code by removing all arguments
and inserting constant operations into the IR code fragment before
executing it.
- Added a missing catch error for gRPC workers' responses.