[DAPHNE-#xyz] Config parameter for sparsity threshold
This commit adds a configuration item to DaphneUserConfig to let the user specify a sparsity threshold. This is arbitrarily set to 25% by default (from the non-configurable previous default of 10%).
Furthermore, all the required changes to read the threshold (from json) to passing it down to the SelectMatrixRepresentationsPass are implemented.
Two more misc comments:
* Added the config param explain_select_matrix_repr to the default UserConfig.json
* Removed createSelectMatrixRepresentationPass from Passes.td as passing arguments is not supported by TableGen (as I found out here [1])

[1](https://github.com/llvm/llvm-project/issues/61301)

Closes #XYZ