[BUGFIX] Inner join running out of memmory
Rootcouse: Due to an intensive call of get() function on frames
the program was running out of memory for values bigger than
10^6 (number of rows of the final result).

Solution: Decouple the logic in order to only call the get()
funcion only once for the insersaction and once for each colum
while composing the result.

Extra: Creation of a hit vector (intersection) that is used
while constructing the result, in order to take advantage of the
temporal locality.