[MINOR] add ccache, lld to CMakeLists.txt
Adds a check for `ccache` and `lld` to the `CMakeLists.txt` and uses
them if available.

`ccache` should help with incremental compilation while developing
DAPHNE.

`lld` reduces the linking time significantly.

Also enables a flag that produces a `compile_commands.json` in the
`build/` directory. This file is used by different development tools and
is an easy way to enable proper LSP functionality (VSCode cpp or clangd
extension or coc-clangd in Neovim).

- https://ccache.dev/
- https://lld.llvm.org/
- https://code.visualstudio.com/docs/cpp/faq-cpp#_how-do-i-get-intellisense-to-work-correctly
- https://marketplace.visualstudio.com/items?itemName=llvm-vs-code-extensions.vscode-clangd
- https://github.com/clangd/coc-clangd