ErrorHandler and KernelDispatchMapping
This commit introduces facilities to unify error handling and improve
the user facing error messages itself.
The ErrorHandler class is used to create new exceptions by passing
required information to the ErorrHandler depending on the type and
source of the exception.
The KernelDispatchMapping provides a way to map a specific kernel call
to a source code line from the DaphneDSL script.

Previously, the following error message was displayed for a failing
lowering pass during compilation:
[error]: MSG

The new error message now includes more useful information:
[error]: HEADER [ MSG ]
     | Source file -> FILE:LINE:COL
     |
LINE | SOURCE
     | ^~~
     |

Additionally, the IR that caused the exception is dumped to disk to help
with debugging and reproducability.

Runtime exceptions exceptions display an error message with the same
format and source information.

Note: KernelDispatchMapping is currently not supported for vectorized or
distributed operations.