Fix leaking file descriptor during test execution
This patch fixes leaking file descriptors during test exectuion. After
the parent process finished reading from the childs pipes, the read ends
of the pipe were not closed. This caused two problems, the call to
`pipe` could fail witth `EMFILE` (errno 24), or cause a deadlock during
test execution. I would guess that happens when the fd remains alive and
is reused which could lead to a scenario were one process writes to the
fd before there is a process reading from it (deadlock-prone).