[MINOR] Improve naming scheme of pooling kernels
This commit introduces the NN c++ namespace for neural network operations and at the same time fixes the problem of pooling ops not found because the naming scheme was wrong. Also fixed an issue with blank space in result type definition in kernels.json causing DAPHNE to fail at runtime.