--statistics flag to track runtime execution
Adds a new CLI command, --statistics, which enables kernel dispatch
level runtime profiling. After execution of the DAPHNE script, the
gathered statistics are displayed to the user as a list of kernel name,
time spent, invocation count, average time per invocation, and the file,
line, and column location in the source file of each operator.

Not tested with distributed pipelines or CUDA pipelines. Follow-up
issue will be created regarding the fixed amount of statistics
currently displayed.