Fix copy constructor chain bug in KernelCatalog
When fetching the `DaphneUserConfig` of the `DaphneContext` with the
`getUserConfig` function, the `DaphneUserConfig` is copy constructed as
it is returned by value. This in turn causes the `KernelKatalog` and
`KernelInfo` to be default copy constructed. Somehow during the
allocation/construction of the internal
`std::unordered_map<std::string, std::vector<KernelInfo>>`, execution
fails in the STL. I'm not sure what exactly is causing the issue here,
the `KernelCatalog` and `KernelInfo` don't seem to be faulty.
For now, returning the `DaphneUserConfig` by reference instead of by
value resolves this problem as this does not copy construct the objects.

To reproduce backtrace run:
`bin/daphne test/api/cli/secondorder/map_3.daphne` before this patch.