[DAPHNE-#770] Refactor getLine() to avoid potential memory leak
According to [1], the use of the C function getLine() will automatically malloc() the required memory if NULL is passed as the first parameter. Since we constantly pass NULL in a loop without ever free()-ing the returned line this constitutes a memory leak. This commit changes the behaviour to always pass the same pointer that is now contained in the File struct. With this the malloc() happens once, the memory is subsequently realloc()-ed if needed and the pointer is passed to free() when closing the file.

[1] https://linux.die.net/man/3/getline