[DAPHNE-#595] Fix crash when using UDF with zero results where one value is needed (#605)
The core issue is that DaphneDSL UDFs without return value internally returned an invalid mlir::Value. Upon doing anything with this value (i.e., when using it in a context where exactly one value is expected, e.g., in expressions or in assignments), the program would crash (e.g., accessing any fields or calling any methods). Now we return nullptr as the result of a UDF with zero return values in visitCallExpr(). This nullptr is detected in valueOrError(), a helper function we anyway call everywhere in the DaphneDSL parser when a single value is needed, and causes an exception to be thrown

The error message is not informative yet ("[error]: While parsing: something was expected to be an mlir::Value, but it was none"), but this will be improved in a follow-up commit.

Added several script-level test cases, which check that DAPHNE does not accept UDFs with zero or more than one return value in places where exactly one value is expected. All of these tests on a UDF with zero return values used to crash DAPHNE with a segfault (#595 was not specific to call expressions). Actually, these are just a few examples, ideally we should have test cases for all uses of expr in the DaphneDSL grammar.

Closes #595.


---------

Co-authored-by: Patrick Damme <patrick.damme@tu-berlin.de>