Implementation of new SparsityOp (#790)
This commit introduces a new SparsityOp that allows users to obtain the compile-time sparsity estimate of a matrix. The function can be called using sparsity(Matrix) in DaphneDSL, which is replaced with the sparsity value through canonicalization if the sparsity is known at compile-time. If the sparsity remains unknown at compile-time, a runtime kernel is called that always returns -1.

This commit does not add any tests. The SparsityOp functionality will be tested as part of #766.