[DAPHNE-#793] Prevent merging transpose into initial SPGEMM
This commit adds a quick fix to not merge transpose operations into sparse matrix multiplication as our CPP kernels can not handle this. The temporary workaround until we support transposition uses a hard coded sparsity threshold as there is no user config available in the canonicalization. The matrix representation selection functionality would also be a suitable measure but is not implemented in a suitable way for this at the moment.