[DAPHNE-#787] DecisionTreeRandomForestTest test result
There is a certain test result in DecisionTreeRandomForestTest that fails on more modern Linux distributions but worked fine on our originally supported Ubuntu 20. This change relaxes the requirement on the test result to let the test suite run through.
The reason for the changed behavior is that we make use of std::uniform_xxx_distribution and the output of this distribution is not only platform but even implementation defined. The behavior of the implementation we use (that comes with Ubuntu's default compiler GCC) changed somewhere around GCC 11 (Ubuntu 20's default is GCC 9). See GitHub issue #787 for a linked discussion on the matter.

Resolves #787