[DAPHNE-#767] Initial HDFS support
This commit adds initial support to read and write files from Hadoop Filesystems in distributed mode. Besides the read, write and distributed functionality, this also contains new configuration options and a new context object to manage the connection information to the distributed filesystem. Finally, this feature requires the installation of more external dependencies. The compilation is therefore optional and can be activated with the --hdfs flag to build.sh.

Closes #767

Co-authored-by: KostasBitsakos <kbitsak@cslab.ece.ntua.gr>
Co-authored-by: Mark Dokter <mark@dokter.cc>