Introduces experimental kernels folder
This commit introduces the `experimental/` folder for kernels. This
folder will contain alternative, experimental implementations of our
kernels. These implementations are using our kernel catalog feature from

While being experimental, the kernels are not tightly coupled to
DAPHNE, e.g, they are not used per default without specifying the
required kernel hint to call the kernel and they are not tested as part
of our test suite. Once an experimental kernel is to be moved to our
default kernels, passing all the tests is required.

Making kernels executable when running DAPHNE while not yet being part
of our kernel library helps the development when improving kernels.

- Dependencies can be used while being experimental without having to
  worry about making them a requirement for all DAPHNE users
- Prototyping and development speed is increased, as only the kernel has
  to be recompiled this is drasticly faster than having to compile the
  `libAllKernels.so` each time
- Provides a playing ground for developers to try out alternative
  implementations
- Makes it easy to benchmark and compare different implementations of
  the same kernel by using different kernel hints

The `gemv/` folder contains a simple example of such an alternative
implementation. It is using AVX2 instructions to implement the SpMV
kernel, making it unsuitable to be the default kernel implementation as
it requires hardware-specific instructions. Additionally, it uses the
LIKWID library to benchmark CPU performance counters (similar to PAPI).
Without bringing these dependency to all DAPHNE users, one can already
test the kernel, compare it with the default, run benchmarks with DAPHNE
using this kernel. For more information on kernel extensions see
https://daphne-eu.github.io/daphne/Extensions/.