[ci] adds clang-format action (#833)
This patch adds an action that runs clang-format on pushes to main and
when opening pull requests. The action runs `clang-format` with version
`18.1.8` on the `src/` and `test/` folders. If the pushed code contains
ill-formatted code, the actions displays the diff.

Also adds the required steps in order to install clang-format-18 in the
Dockerfile.