VectorizedPipelineOp deduplication bug (#697)
To find duplicates, the canonicalize method of the VectorizedPipelineOp
uses a std::vector to store the VectorSplits of its inputs. In the
actual loop, the canonicalize method didn't use this std::vector, but
was still fetching the splits from the VectorizedPipelineOp itself. This
patch fixes this mistake. Now inputs that are used multiple times and
have multiple splits are correctly considered for each usage.

Closes #697