[build] set -fno-omit-frame-pointer
This patch sets the `-fno-omit-frame-pointer` for non release builds.
This is useful for a few things:
- makes stack traces more reliable and complete
- enables better profiling with tools like perf
- better visibility into call chains during performance analysis

It has a small impact on performance as an extra register needs to be
used, e.g., the `rbp` register on x86_64 to store the frame pointer. It
also incurs a small of setup and teardown of stack frames. Because of
these reasons, this is currently only set for non release builds.