[kernels] manually unlock mutex before cv notify
This patch adds calls to `std::unique_lock<Mutex>::unlock` to remove
potential cases where a thread can be woken up only to block again.

"The notifying thread does not need to hold the lock on the same mutex
as the one held by the waiting thread(s); in fact doing so is a
pessimization, since the notified thread would immediately block again,
waiting for the notifying thread to release the lock."

https://en.cppreference.com/w/cpp/thread/condition_variable
https://en.cppreference.com/w/cpp/thread/condition_variable/notify_one

Uses consistent name for all instances of `std::unique_lock` in
TaskQueue.h.