[kernels] --vec thread count causes oob access
A wrongly set thread count in the vectorized pipeline kernel causes an
out-of-bounds access on a std::vector when run with `_queueMode = 1`.

This is happenening because the MTWrapper sets the _numCPPThreads count
to std::thread::hardware_concurrency, which returns the number of
_logical_ cores.
https://github.com/daphne-eu/daphne/blob/cfd34a484fef73186c84c83f89f5bd1a3397458b/src/runtime/local/vectorized/MTWrapper.h#L192

Meanwhile, the get_topology function uses hwloc to query the node
topology and puts the socket id per _physical_ core in a vector, said
vector is passed to the WorkerCPU constructor and used in the
WorkerCPU::run function.

https://github.com/daphne-eu/daphne/blob/cfd34a484fef73186c84c83f89f5bd1a3397458b/src/runtime/local/vectorized/WorkerCPU.h#L61

In this function, the vector is accessed with operator[] using the
_threadID. When setting the amount of threads to the amount of _logical_
cores, and the vector of socket ids only containing the ids for
_physical_ cores, this causes an out-of-bounds access. This patch caps
the number of threads used to the amount of _phyiscal_ cores.