[codegen] Refactor of codegen passes [#889]
This PR majorly reworks codegen for AllAgg* and EwOps as well as add
lowering for TransposeOp and Row/ColAgg*. All of these passes are added
to the optional MLIR codegen pipeline that can be enabled using the
--mlir-codegen flag and offer alternative lowering of these operations
to MLIR rather than calls to precompiled C++ kernels. Currently, they
only support DenseMatrix with dimensions that are known at compile-time
and any value type (except Booleans).

Except for IdxMin, IdxMax which are directly lowered to affine loops and
TransposeOp which lowers to a named linalg op all passes make use of
linalg GenericOps which are then lowered to affine loops in a later pass
in the codegen pipeline. They convert the input DenseMatrix to a MemRef
and create a new MemRef for the output that is converted into a
DenseMatrix.

Changes:
- Add codegen for AllAgg*Op, Row/ColAgg*Op, Ew*Op and TransposeOp (see
  below for details)
- Added passes to TableGen files and codegen pipeline
- Added script level test cases / MLIR test cases (using FileCheck)
- Replaced old tests Renamed some old test scripts for EwOps for better
  organization
- Edited fusion.mlir test to lower Linalg to affine loops before
  applying fusion pass
- Added Canonicalization passes for floor, ceil, round that removes the
  respective ops when input type is an integer (this also simplifies
  codegen)
- Added some necessary instantiations in kernels.json
- Restored alphabetic sorting of codegen passes in ir/daphneir/Passes.h

Ops with new codegen:
- AllAgg*Op Sum, Min, Max
- Row/ColAgg*Op Sum, Min, Max, IdxMin, IdxMax
- Ew*Op Unary (scalar/matrix): Abs, Sqrt, Exp, Ln, Sin, Cos, Floor,
  Ceil, Round Binary (scalar-scalar/matrix-matrix/matrix-scalar
  broadcasting): Add, Sub, Mul, Div, Pow, Max, Min
- TransposeOp

Known limitations are listed in the PR description [#889]

Co-authored-by: philipportner ortner@tu-berlin.de