[codegen] broadcasting of rhs in ewBinary codegen
This PR changes the existing ewBinaryOp codegen to broadcast rhs if possible.
So far, the ewBinary codegen only accepted scalars, matrices of equal
shape or combinations of lhs being a matrix and rhs being either a
scalar or singleton. Now rhs can also be given as a matching (equal to
lhs in one dimension and size 1 in the other) row or column vector.

- changed index map for rhs in existing linalg genericOp to allow
  broadcasting of row/column vectors
- added shape checks to ensure matching dimensions when rhs is broadcast

closes #920