Added string support for the SQL parser
The SSB Benchmark uses strings for comparisions, which
this implementation supports, by useing the same
infrastructure the comparisions used before.
This does not include type checking, such that this
general implementation might make it possible for
unexpected errors. For instance the SQL Grammar allows
a multiplication between a String and an Integer.
For handeling of such an Error I trust the DAPHNE
infrastructure.