Added Test Cases for new SQL Parser functionalities
Test for String and Between support where added.
The test writing exposed an already an existing issue for the current implementation,
which lead to a quick fix.
Issue #203 is a persistent issue for the sql parser,
for instance in the where clause it is legal to: f.a < 1 however it is illegal to 1 > f.a,
because the parser can't diferentiate between the two. This had an implication for between,
since in the code gen the objects where switched and the operator state the same. This got
now rectified, such that LessEqual and GreaterEqual Ops get used instead of just LessEqual.