[kernel] CUDAPass fixes output size calculation
This patch fixes the output size calculation in the `fitsInMemory`.
Before, for unknown shapes, the number of rows / cols could be `-1`, and
thus cause wrong estimates for the output size. When calculating the
input size, this is already handled and unknown sizes are set to `1`.
This now also happens for the output size calculation.